requirements.txt
pyserial>=3.5
hass-mqtt-things==0.5.9
